requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForms==3.0.1
